docs: add markdown cross references to truth.bzl
This uses a markdown trick to make it easy to write a type name that
should show up in code-font _and_ become a link to somewhere. When
given `[foo]`, markdown will automatically look for a url reference
named "foo", which can be provided elsewhere on the page, e.g. by
logic to create generated docs.

This means, as part of generating doc, a "crossrefs.md" file can be appended
to allow primitive crossrefs.

Also some various doc fixups:
* Linkify types mentioned in formal descriptions of
  interfaces/signatures.
* Linkify the first mention of any types in prose text; code font
  subsequent mentions.
* Clarify the types of some type descriptions.
* Various grammar/spelling fixes
* Standarizes on `argname: type; description` formatting.
* Fixes example: remove `actions` from test function signature.

PiperOrigin-RevId: 517214297