feat: Add CollectionSubject.offset
This is a relatively thin accessor. The main advantage over directly
accessing the underlying value is the additional context information is propagated.

The user must manually pass a `factory` arg to construct the wrapped value for two reasons:
  * A collection doesn't know the types of elements it contains.
  * CollectionSubject can't depend on StrSubject because it would cause a circular
    dependency (StrSubject.split requires CollectionSubject), and finding a
    solution for this looks non-trivial.

A factory arg is necessary anyways to support custom subjects.

Also exposes StrSubject as `subjects.str` (as done with several other subjects)

Closes #41

PiperOrigin-RevId: 540009713