docs: add human-friendly changelog
This adds a changelog in keepachanglog.com-style format so that it's
easier to see what features and changes are present in a release.

The log only goes back to 0.0.5, since that's the first real release -- one
that wasn't part of testing of releases and cleanup as part of initial
releasing.

PiperOrigin-RevId: 546784896