chore: upgrade (dev only dependency) rules_python 0.22.0 -> 0.27.0
This removes a hack to make the dev-only register_toolchains() call
not break the production configuration. This hack is no longer necessary
as of rules_python 0.23.0 and higher. Version 0.27.0 is chosen because
the latest release has some code for building Sphinx documentation
I'd like to experiment with re-using.

Coincidentally, this should also fix an an issue with the Bazel@HEAD
downstream CI pipeline running tests for Mac/Windows that should be
skipped, but fail due to very early MODULE-phase errors. With this
change, it's able to get past that failure and run enough that the
Bazel@HEAD downstream CI should be happy. Note those tests aren't
gauranteed to continue to work on those platforms, they just happen to.

Fixes https://github.com/bazelbuild/rules_testing/issues/76