Add support for provider maps to rules_testing
Currently, If you are testing a custom provider, the API requires you to write:
    env.expect.that_target(targets.foo).provider(FooInfo, factory=FooFactory)...
    env.expect.that_target(targets.foo).provider(BarInfo, factory=BarFactory)...

This can get very tedious, and more importantly, is not very safe, since you can write:
    env.expect.that_target(targets.foo).provider(FooInfo, factory=FooFactory)...
    env.expect.that_target(targets.foo).provider(BarInfo, factory=FooFactory)...

Additionally, custom types are always rendered as "<provider>".

To solve this, we add the ability to directly specify a list of factories for custom types in your test.

analysis_test(
    ...,
    provider_factories = [struct(type = FooInfo, name = "FooInfo", factory = FooFactory)]
)

PiperOrigin-RevId: 607174706